% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Consolidado}
\alias{Consolidado}
\title{Consolidado}
\format{
Un data frame con cinco variables:
\describe{
\item{\code{Variable}}{Abc defg hijk.}
\item{\code{YEAR}}{Abc defg hijk.}
\item{\code{SEMESTRE}}{Abc defg hijk.}
\item{\code{Clase}}{Abc defg hijk.}
\item{\code{Total}}{Abc defg hijk.}
}
}
\usage{
Consolidado
}
\description{
Esta es un oración de ejemplo proporcionada por ABC.
}
\examples{
head(Consolidado)
}
\keyword{datasets}
