% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejConsolidadoGrad}
\alias{ejConsolidadoGrad}
\title{Consolidado de Graduados}
\format{
Un data frame (\emph{data.frame, tbl_df o tbl}) con
775 filas
y 5 columnas:
'Variable', 'YEAR', 'SEMESTRE', 'Clase', 'Total'.
}
\source{
Para obtener más detalle de los metadatos consulte \href{https://estadisticaun.github.io/DabiertosUNAL/}{aquí}.
}
\usage{
ejConsolidadoGrad
}
\description{
Consolidado de la población de graduados de los programas académicos en la
Universidad Nacional de Colombia, cuenta con la información histórica desde
el 2009-I al 2021-I.
}
\examples{
\dontshow{if (require("dplyr")) withAutoprint(\{ # examplesIf}
# library(dplyr)
head(ejConsolidadoGrad)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
