% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejGraduados}
\alias{ejGraduados}
\title{Muestra de Microdatos de Graduados}
\format{
Un data frame (\emph{data.frame, tbl_df o tbl}) con 22825
filas y 53 columnas:
\code{YEAR}, \code{SEMESTRE}, \code{TIPO_NIVEL}, \code{NIVEL}, \code{DEP_NAC}, \code{COD_DEP_NAC}, \code{CIU_NAC}, \code{COD_CIU_NAC}, \code{LON_CIU_NAC}, \code{LAT_CIU_NAC}, \code{DEP_PROC}, \code{COD_DEP_PROC}, \code{CIU_PROC}, \code{COD_CIU_PROC}, \code{LON_CIU_PROC}, \code{LAT_CIU_PROC}, \code{CODS_NAC}, \code{CODN_NAC}, \code{NACIONALIDAD}, \code{EDAD_MOD}, \code{CAT_EDAD}, \code{SEXO}, \code{ESTRATO_ORIG}, \code{ESTRATO}, \code{TIPO_COL}, \code{PBM_ORIG}, \code{PBM}, \code{MAT_PVEZ}, \code{DISCAPACIDAD}, \code{TIPO_DISC}, \code{SNIES_SEDE_ADM}, \code{SEDE_NOMBRE_ADM}, \code{SNIES_SEDE_MAT}, \code{SEDE_NOMBRE_MAT}, \code{ADM_PEAMA_ANDINA}, \code{MOD_ADM}, \code{TIPO_ADM}, \code{PAES}, \code{PEAMA}, \code{MOV_PEAMA}, \code{CONVENIO}, \code{TIP_CONVENIO}, \code{SNIESU_CONVENIO}, \code{U_CONVENIO}, \code{FACULTAD}, \code{FACULTAD_S}, \code{SNIES_PROGRA}, \code{PROGRAMA}, \code{PROGRAMA_S}, \code{AREAC_SNIES}, \code{CA_CINE}, \code{CD_CINE}, \code{AREA_CINE}.
}
\source{
Para obtener más detalle de los metadatos consulte \href{https://n9.cl/8qk3y}{aquí}.
}
\usage{
ejGraduados
}
\description{
Muestra de microdatos de la población de graduados de los programas académicos
en la Universidad Nacional de Colombia, se cuenta con una muestra de la información
disponible (\emph{del 2019-I al 2021-I}). Dicho dataset será usado en los ejemplos
de las funciones \code{Tabla.General()} y \code{Plot.Treemap()}.
}
\examples{
ejGraduados[1:5, 1:10]
}
\keyword{datasets}
