% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejMiniAspirantesPre}
\alias{ejMiniAspirantesPre}
\title{Mini Microdatos de Aspirantes a Pregrado}
\format{
Un data frame (\emph{data.frame, tbl_df o tbl}) con
30.050
filas y 6 columnas:
'Serie', 'PTOTAL', 'TIPO_INS', 'INS_SEDE_NOMBRE', 'ADM_SEDE_NOMBRE', 'FACULTAD'.
}
\source{
Para obtener más detalle de los metadatos consulte \href{https://estadisticaun.github.io/DabiertosUNAL/}{aquí}.
}
\usage{
ejMiniAspirantesPre
}
\description{
Muestra de los microdatos (\emph{únicamente estudiantes de pregrado}) de la población
de aspirantes a cursar estudios de pregrado o postgrado en la Universidad
Nacional de Colombia inscritos a través de convocatoria pública de manera regular
o por medio de los programas de admisión especial existentes. Dicho dataset
será usado en los ejemplos de la función \code{Plot.Boxplot()}.
}
\examples{
\dontshow{if (require("dplyr")) withAutoprint(\{ # examplesIf}
# library(dplyr)
head(ejMiniAspirantesPre)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
