% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejMiniConsolidadoAsp}
\alias{ejMiniConsolidadoAsp}
\title{Mini Consolidado de Aspirantes}
\format{
Un data frame (\emph{data.frame, tbl_df o tbl}) con
308
filas y 5 columnas:
'Variable', 'YEAR', 'SEMESTRE', 'Clase', 'Total'.
}
\source{
Para obtener más detalle de los metadatos consulte \href{https://estadisticaun.github.io/DabiertosUNAL/}{aquí}.
}
\usage{
ejMiniConsolidadoAsp
}
\description{
Mini consolidado de la población de aspirantes a cursar estudios de pregrado
o postgrado en la Universidad Nacional de Colombia inscritos a través de
convocatoria pública de manera regular o por medio de los programas de admisión
especial existentes -\emph{sólo pregrado}-. Dicho dataset será usado en los ejemplos
de la función \code{Plot.Drilldown()}.
}
\examples{
\dontshow{if (require("dplyr")) withAutoprint(\{ # examplesIf}
# library(dplyr)
head(ejMiniConsolidadoAsp)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
