% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Graduados}
\alias{Graduados}
\title{Graduados}
\format{
Un data frame con 3813 filas y 53 columnas.
}
\source{
Para obtener más detalles de los metadatos, consulte \url{http://estadisticas.unal.edu.co/menu-principal/cifras-generales/metadatos/cifras-generales/}
}
\usage{
Graduados
}
\description{
Microdatos de los estudiantes graduados concerniente al periodo 2020-II de la
Universidad Nacional de Colombia.
}
\examples{
Graduados[1:5, 1:10]
}
\keyword{datasets}
