% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Create a package skeleton}
\usage{
hello(a)
}
\arguments{
\item{a}{character: name for your new package}
}
\value{
Abc def, \code{jkm}.
}
\description{
Create a package skeleton
}
