% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleExcel.R
\name{read_example}
\alias{read_example}
\title{Obtenga la ruta de los archivos de Excel para la función \code{Agregar()}}
\usage{
read_example(ruta = NULL)
}
\arguments{
\item{ruta}{Nombre del archivo. Si es \code{NULL}, se enumerarán los archivos de
ejemplo.}
}
\description{
UnalR viene con algunos archivos de ejemplo en su directorio \code{inst/extdata}.
Esta función facilita el acceso a ellos.
}
