% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SaberPro}
\alias{SaberPro}
\title{SaberPro}
\format{
Un data frame con cinco variables:
\describe{
\item{\code{Variable}}{Abc defg hijk.}
\item{\code{YEAR}}{Abc defg hijk.}
\item{\code{Clase}}{Abc defg hijk.}
\item{\code{n}}{Abc defg hijk.}
\item{\code{Componente}}{Abc defg hijk.}
\item{\code{Total}}{Abc defg hijk.}
\item{\code{desv}}{Abc defg hijk.}
}
}
\source{
Para obtener más detalles, consulte \url{https://www.icfes.gov.co/investigadores-y-estudiantes-posgrado/acceso-a-bases-de-datos}
}
\usage{
SaberPro
}
\description{
Esta es un oración de ejemplo proporcionada por ABC.
}
\examples{
head(SaberPro)
}
\keyword{datasets}
