% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StaticPlot.R
\name{StaticPlot}
\alias{StaticPlot}
\title{Guarde y presente un widget HTML renderizado como una imagen estática}
\usage{
StaticPlot(HTML_Widget, Height = 500, PrimeraVez = FALSE, ...)
}
\arguments{
\item{HTML_Widget}{Widget HTML a ser mostrado de forma estática.}

\item{Height}{Altura de la imagen estática a retornar.}

\item{PrimeraVez}{Si es \code{FALSE} (valor predeterminado) no se instalará \code{PhantomJS}.
Éste es necesario instalarlo una única vez, por lo cual si es la primera vez
que corre la función deberá indicar el argumento con el valor \code{TRUE}, después
de esto omita este argumento y deje su valor por defecto.}

\item{...}{Otros parámetros concernientes a la función \link[webshot:webshot]{webshot()},
sin considerar los ya usados dentro de la función (\code{url}, \code{file}, \code{delay}, \code{zoom} y \code{vheight}).}
}
\value{
Una imagen estática.
}
\description{
Esta función permite representar un widget HTML como un objeto ráster
(\emph{imagen de mapa de bits}), útil para reproducir gráficos interactivos en
archivos estáticos como un \code{.pdf} generado por \verb{R Markdown}. Esta función usa
internamente los paquetes \code{webshot}, \code{htmlwidgets}, \code{png} y \code{grid} para poder
llevar a cabo su propósito.
}
\details{
No es necesario especificar el número del factor de zoom. El factor de zoom por
defecto es 5, el cual dará como resultado cinco veces más de píxeles vertical y
horizontalmente. Este valor fue seleccionado debido a que es el óptimo, un valor
mayor ocasiona un tiempo de ejecución excesivamente alto y poca ganancia en cuanto
a calidad, con un valor menor se tiene una pérdida considerable de calidad.

Si se especifican tanto el ancho como el alto, es probable que la imagen se distorsione.
Por lo tanto, el único argumento variable será la altura de la imagen, dejando el
ancho como un argumento adaptativo dependiendo del ancho disponible.

El tiempo de espera antes de tomar una captura de pantalla, en segundos, es de
\eqn{1s}. Este valor debido a que se necesita un retraso mayor para que los grafico
generados por \code{Highcharter} se muestren correctamente.
}
\examples{
\dontrun{
col <- c("#29ABE2", "#8CC63F", "#CC241D",
         "#FF3673", "#0071BC", "#F15A24",
         "#FBB03B", "#93278F", "#8A381A")
A <- Plot.Series(datos = Consolidado, categoria = "SEDE_NOMBRE_ADM", colores = col)
StaticPlot(A)

ano <- 2020; semestre <- 2
col <- c("#116BEE", "#E62272")
B <- Plot.Torta(datos     = Consolidado,
                categoria = "SEXO",
                ano       = ano,
                periodo   = semestre,
                colores   = col,
                titulo    = "DISTRIBUCI\u00d3N DE GRADUADOS POR SEXO",
                libreria  = "highcharter")
StaticPlot(B)
}

}
