% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejConsolidadoSaberPro2019}
\alias{ejConsolidadoSaberPro2019}
\title{Consolidado Saber Pro 2019}
\format{
Un data frame (\emph{data.frame, tbl_df o tbl}) con
42
filas y 7 columnas:
'Variable', 'YEAR', 'Clase', 'n', 'Componente', 'Total', 'desv'.
}
\source{
Para obtener más detalle de los metadatos consulte \href{https://estadisticaun.github.io/DabiertosUNAL/}{aquí}.
}
\usage{
ejConsolidadoSaberPro2019
}
\description{
Consolidado de ejemplo de la prueba Saber Pro (\emph{antes llamada ECAES}) del año
2019. El cual se utiliza para los ejemplos de la función \code{Tabla.SaberPro()}.
}
\examples{
\dontshow{if (require("dplyr")) withAutoprint(\{ # examplesIf}
# library(dplyr)
head(ejConsolidadoSaberPro2019)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
