% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejConsolidadoSaberPro2019}
\alias{ejConsolidadoSaberPro2019}
\title{Consolidado Saber Pro 2019}
\format{
Un data frame (\emph{data.frame, tbl_df o tbl}) con \verb{r nrow(ejConsolidadoSaberPro2019)}
filas y \verb{r ncol(ejConsolidadoSaberPro2019)} columnas:
\verb{r paste0("``", colnames(ejConsolidadoSaberPro2019), "``", collapse = ", ")}.
}
\source{
Para obtener más detalle de los metadatos consulte \href{https://n9.cl/c08al}{aquí}.
}
\usage{
ejConsolidadoSaberPro2019
}
\description{
Consolidado de ejemplo de la prueba Saber Pro (\emph{antes llamada ECAES}) del año
2019. El cual se utiliza para los ejemplos de la función \code{Tabla.SaberPro()}.
}
\examples{
if (require("dplyr")) {
  head(ejConsolidadoSaberPro2019)
}
}
\keyword{datasets}
