% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejGraduados}
\alias{ejGraduados}
\title{Muestra de Microdatos de Graduados}
\format{
Un data frame (\emph{data.frame, tbl_df o tbl}) con
27.830 filas y
53 columnas:
'YEAR', 'SEMESTRE', 'TIPO_NIVEL', 'NIVEL', 'DEP_NAC', 'COD_DEP_NAC', 'CIU_NAC', 'COD_CIU_NAC', 'LON_CIU_NAC', 'LAT_CIU_NAC', 'DEP_PROC', 'COD_DEP_PROC', 'CIU_PROC', 'COD_CIU_PROC', 'LON_CIU_PROC', 'LAT_CIU_PROC', 'CODS_NAC', 'CODN_NAC', 'NACIONALIDAD', 'EDAD_MOD', 'CAT_EDAD', 'SEXO', 'ESTRATO_ORIG', 'ESTRATO', 'TIPO_COL', 'PBM_ORIG', 'PBM', 'MAT_PVEZ', 'DISCAPACIDAD', 'TIPO_DISC', 'SNIES_SEDE_ADM', 'SEDE_NOMBRE_ADM', 'SNIES_SEDE_MAT', 'SEDE_NOMBRE_MAT', 'ADM_PEAMA_ANDINA', 'MOD_ADM', 'TIPO_ADM', 'PAES', 'PEAMA', 'MOV_PEAMA', 'CONVENIO', 'TIP_CONVENIO', 'SNIESU_CONVENIO', 'U_CONVENIO', 'FACULTAD', 'FACULTAD_S', 'SNIES_PROGRA', 'PROGRAMA', 'PROGRAMA_S', 'AREAC_SNIES', 'CA_CINE', 'CD_CINE', 'AREA_CINE'.
}
\source{
Para obtener más detalle de los metadatos consulte \href{https://estadisticaun.github.io/DabiertosUNAL/}{aquí}.
}
\usage{
ejGraduados
}
\description{
Muestra de microdatos de la población de graduados de los programas académicos
en la Universidad Nacional de Colombia, se cuenta con una muestra de la información
disponible (\emph{del 2019-I al 2021-I}). Dicho dataset será usado en los ejemplos
de las funciones \code{Tabla.General()} y \code{Plot.Treemap()}.
}
\examples{
\dontshow{if (require("dplyr")) withAutoprint(\{ # examplesIf}
# library(dplyr)
ejGraduados[1:5, 1:10]
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
