% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejGraduados}
\alias{ejGraduados}
\title{Muestra de Microdatos de Graduados}
\format{
Un data frame (\emph{data.frame, tbl_df o tbl}) con \verb{r nrow(ejGraduados)}
filas y \verb{r ncol(ejGraduados)} columnas:
\verb{r paste0("``", colnames(ejGraduados), "``", collapse = ", ")}.
}
\source{
Para obtener más detalle de los metadatos consulte \href{https://n9.cl/8qk3y}{aquí}.
}
\usage{
ejGraduados
}
\description{
Muestra de microdatos de la población de graduados de los programas académicos
en la Universidad Nacional de Colombia, se cuenta con una muestra de la información
disponible (\emph{del 2019-I al 2021-I}). Dicho dataset será usado en los ejemplos
de las funciones \code{Tabla.General()} y \code{Plot.Treemap()}.
}
\examples{
ejGraduados[1:5, 1:10]
}
\keyword{datasets}
