% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejMiniAspirantesPre}
\alias{ejMiniAspirantesPre}
\title{Mini Microdatos de Aspirantes a Pregrado}
\format{
Un data frame (\emph{data.frame, tbl_df o tbl}) con \verb{r nrow(ejMiniAspirantesPre)}
filas y \verb{r ncol(ejMiniAspirantesPre)} columnas:
\verb{r paste0("``", colnames(ejMiniAspirantesPre), "``", collapse = ", ")}.
}
\source{
Para obtener más detalle de los metadatos consulte \href{https://n9.cl/4hoq}{aquí}.
}
\usage{
ejMiniAspirantesPre
}
\description{
Muestra de los microdatos (\emph{únicamente estudiantes de pregrado}) de la población
de aspirantes a cursar estudios de pregrado o postgrado en la universidad
nacional de Colombia inscritos a través de convocatoria pública de manera regular
o por medio de los programas de admisión especial existentes. Dicho dataset
será usado en los ejemplos de la función \code{Plot.Boxplot()}.
}
\examples{
if (require("dplyr")) {
  head(ejMiniAspirantesPre)
}
}
\keyword{datasets}
