% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejMiniConsolidadoAsp}
\alias{ejMiniConsolidadoAsp}
\title{Mini Consolidado de Aspirantes}
\format{
Un data frame (\emph{data.frame, tbl_df o tbl}) con 297
filas y 5 columnas:
\code{Variable}, \code{YEAR}, \code{SEMESTRE}, \code{Clase}, \code{Total}.
}
\source{
Para obtener más detalle de los metadatos consulte \href{https://n9.cl/4hoq}{aquí}.
}
\usage{
ejMiniConsolidadoAsp
}
\description{
Mini consolidado de la población de aspirantes a cursar estudios de pregrado
o postgrado en la universidad nacional de Colombia inscritos a través de
convocatoria pública de manera regular o por medio de los programas de admisión
especial existentes -\emph{sólo pregrado}-. Dicho dataset será usado en los ejemplos
de la función \code{Plot.Drilldown()}.
}
\examples{
if (require("dplyr")) {
  head(ejMiniConsolidadoAsp)
}
}
\keyword{datasets}
