% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ejSaberPro2020}
\alias{ejSaberPro2020}
\title{Microdatos Saber Pro 2020}
\format{
Un data frame (\emph{data.frame, tbl_df o tbl}) con \verb{r nrow(ejSaberPro2020)}
filas y \verb{r ncol(ejSaberPro2020)} columnas:
\verb{r paste0("``", colnames(ejSaberPro2020), "``", collapse = ", ")}.
}
\source{
Para obtener más detalle de los metadatos consulte \href{https://n9.cl/c08al}{aquí}.
}
\usage{
ejSaberPro2020
}
\description{
Microdatos de los resultados de la prueba Saber Pro del año 2020, obtenidos
desde la página del ICFES, los cuales serán usados para los ejemplos de las
funciones \code{Plot.Boxplot()} y \code{Plot.Radar()}.
}
\examples{
ejSaberPro2020[1:5, 1:10]
}
\keyword{datasets}
