% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Apiladas.R
\name{Plot.Apiladas}
\alias{Plot.Apiladas}
\title{Cree un gráfico de barras apiladas dinámico y flexible}
\usage{
Plot.Apiladas(
  datos,
  categoria,
  anos,
  periodo,
  colores,
  titulo = "",
  addPeriodo = TRUE,
  libreria = c("highcharter", "plotly"),
  estilo = NULL,
  estatico = FALSE
)
}
\arguments{
\item{datos}{Un data frame, no un vector numérico.}

\item{categoria}{Igual uso que en \code{\link[=Plot.Series]{Plot.Series()}}}

\item{anos}{Igual uso que en \code{\link[=Plot.Torta]{Plot.Torta()}}}

\item{periodo}{Igual uso que en \code{\link[=Plot.Torta]{Plot.Torta()}}}

\item{colores}{Igual uso que en \code{\link[=Plot.Series]{Plot.Series()}}}

\item{titulo}{Igual uso que en \code{\link[=Plot.Series]{Plot.Series()}}}

\item{addPeriodo}{Igual uso que en \code{\link[=Plot.Torta]{Plot.Torta()}}}

\item{libreria}{Igual uso que en \code{\link[=Plot.Torta]{Plot.Torta()}}}

\item{estilo}{Lista compuesta por varios parámetros, los cuales van a ser usados
para graficar las barras apiladas y cuyo objetivo es personalizar pequeños
detalles de éste.
\itemize{
\item \code{hc.Tema}, \code{hc.Credits}, \code{gg.Tema}, \code{gg.Legend} y \code{gg.Texto}: Igual uso
que en \code{\link[=Plot.Series]{Plot.Series()}}
\item \code{LegendTitle}: Cadena de caracteres indicado un título para la leyenda
(\emph{diferentes niveles del argumento \code{varPrincipal}}).
\item \code{ply.LegendPosition}: Igual uso que en \code{\link[=Plot.Series]{Plot.Series()}}
\item \code{ply.Credits}: Igual uso que en \code{\link[=Plot.Series]{Plot.Series()}}
\item \code{gg.Bar}: Igual uso que en \code{\link[=Plot.Barras]{Plot.Barras()}}
}}

\item{estatico}{Igual uso que en \code{\link[=Plot.Series]{Plot.Series()}}}
}
\value{
Retorna el diagrama de barras apiladas (\emph{objeto widget de HTML}) creado. La
clase del objeto retornado será un "htmlwidget" y adicionalmente pertenecerá
a la clase "highchart".
}
\description{
Esta función proporciona excelentes herramientas y opciones para la visualización
de un gráfico de barras apiladas con el objetivo de mostrar el tamaño relativo
(\emph{como porcentaje}) de una variable de categorías, subdivididas por colores en
función de un subgrupo. Dicha gráfica se va a representar usando la librería
\code{Highcharter}, la cual usa internamente \code{JavaScript}.
}
\examples{
Msj <- "Se considera \u00fanicamente los valores obtenidos en el primer periodo acad\u00e9mico de cada a\u00f1o."
Plot.Apiladas(
  datos     = ejConsolidadoGrad,
  categoria = "NIVEL",      # Pruebe también con -> unique(ejConsolidadoGrad$Variable)
  anos      = c(2018:2020),
  periodo   = 1,
  colores   = c("#FFA700", "#C10AA1", "#01CDFE", "#00FF44", "#FF0040"),
  titulo    = "BARRAS APILADAS EN FUNCI\u00d3N DEL NIVEL ACAD\u00c9MICO Y EL A\u00d1O",
  estilo    = list(LegendTitle = "NIVEL ACAD\u00c9MICO:", hc.Tema = 4, hc.Credits = Msj)
)
Plot.Apiladas(
  datos     = ejConsolidadoGrad,
  categoria = "AREAC_SNIES",
  anos      = c(2018:2020),
  periodo   = 1,
  colores   = c("#D2D4DC", "#FF8ABF", "#945BC2", "#D11879","#FF7F7F", "#FFA568", "#9CFF86", "#89D8FF"),
  titulo    = "BARRAS APILADAS EN FUNCI\u00d3N DEL \u00c1REA DEL SNIES",
  libreria  = "plotly",
  estilo    = list(LegendTitle = "NIVEL ACAD\u00c9MICO:",
                   ply.Credits = list(x = 0.45, y = 1.1, text = Msj),
                   ply.LegendPosition = list(x = 0.04, y = -0.3, orientation = "h")
  )
)
# Ejemplo usando el caso estático (ggplot2)
Plot.Apiladas(
  datos     = ejConsolidadoGrad,
  categoria = "NIVEL",
  anos      = c(2019:2021),
  periodo   = 1,
  colores   = c("#FFA700", "#C10AA1", "#01CDFE", "#00FF44", "#FF0040"),
  titulo    = "BARRAS APILADAS EN FUNCI\u00d3N DEL NIVEL ACAD\u00c9MICO Y EL A\u00d1O",
  estatico  = TRUE,
  estilo    = list(
    LegendTitle = "NIVEL ACAD\u00c9MICO:", gg.Tema = 8,
    gg.Legend = list(legend.position = "right", legend.direction = "vertical"),
    gg.Bar    = list(width = 0.6, color = "#000000"),
    gg.Texto  = list(subtitle = Msj,
                    caption  = "Informaci\u00f3n Disponible desde 2009-1",
                    tag      = "Figura 1. Graduados"
    )
  )
)

}
