% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Series.R
\name{Plot.Series}
\alias{Plot.Series}
\title{Cree una serie de tiempo dinámica y flexible con tres diferentes paquetes}
\usage{
Plot.Series(
  datos,
  categoria,
  colores,
  titulo = "",
  labelX = "Periodo",
  labelY = "",
  libreria = c("highcharter", "plotly", "dygraphs"),
  estilo = NULL
)
}
\arguments{
\item{datos}{Un data frame, no un objeto clase serie de tiempo o vector numérico.}

\item{categoria}{Una variable categórica dentro del data frame ingresado en \code{datos}.}

\item{colores}{Cadena de caracteres indicando los colores con los cuales se deben
colorear cada una de las series correspondiente a cada nivel del parámetro
\code{categoria}. Si no se introduce algún vector se usará la paleta \code{rainbow} por
defecto.}

\item{titulo}{Cadena de caracteres indicando el título principal del plot.}

\item{labelX}{Cadena de caracteres indicando la etiqueta del eje \code{X}. Por defecto
se emplea el rótulo "Periodo".}

\item{labelY}{Cadena de caracteres indicando la etiqueta del eje \code{Y}.}

\item{libreria}{Cadena de caracteres que indica el paquete con el cual se realizará
la serie. Los valores permitidos son \code{"highcharter"} (valor predeterminado),
\code{"plotly"} o \code{"dygraphs"}. Los valores se emparejarán parcialmente.}

\item{estilo}{Lista compuesta por varios parámetros, los cuales van a ser usados
de acuerdo con la librería especificada para graficar la serie y cuyo objetivo
es personalizar pequeños detalles de ésta.
\itemize{
\item \code{LegendTitle}: Cadena de caracteres indicado un título para la legenda
(\emph{diferentes niveles del parámetro \code{categorias}}). Se utilizará tanto
en el paquete \code{Highcharter} como en \code{Plotly}.
\item \code{hc.Tema}: Modifica el tema con el cual se creará la serie. Los posibles
valores son un número entero entre \eqn{[1, 10]} (\emph{el tema por defecto
es el número 5, \code{hc_theme_flat()}}), el cual hace referencia a diferentes
temas disponibles en dicha librería.
\item \code{hc.Slider}: Si es \code{TRUE} agrega un deslizador/navegador dinámico en la parte
inferior de la serie. Proporciona herramientas para acercar y alejar partes
de la serie, así como para desplazarse por el conjunto de datos. El valor por
defecto es \code{FALSE}.
\item \code{hc.BoxInfo}: Si es \code{TRUE} (valor predeterminado) la información concerniente
a cada punto se visualiza conjuntamente en un cuadro, o de forma individual
(\code{FALSE}) al pasar el cursor.
\item \code{hc.Credits}: Cadena de caracteres indicando un subtítulo o etiqueta de
créditos debajo del título principal.
\item \code{ply.LegendPosition}: Lista que específica la posición y orientación de la
legenda. Los valores por defecto la ubican centrada verticalmente a la
derecha del plot, es decir, \code{c(x = 1, y = 0.5, orientation = "v")}.
\item \code{ply.Interaction}: Cadena de caracteres que determina el modo de las
interacciones de desplazamiento. Los valores permitidos son "x unified"
(valor predeterminado), "y unified", "closest", "x", "y" y FALSE.
\item \code{ply.Credits}: Lista que específica la posición y texto para añadir un
subtítulo o etiqueta de créditos a la serie principal, por ejemplo,
\code{c(x = 0.2, y = 1, text = "https://...")}.
\item \code{dyg.LegendWidth}: Un número que indica el ancho (\emph{en píxeles}) que ocupará
la leyenda. El valor por defecto es 250.
\item \code{dyg.Resaltar}: Si es \code{FALSE} (valor predeterminado) no se resaltará la serie
en que se sitúa el cursor.
}}
}
\value{
Retorna la serie (\emph{objeto widget de HTML}) creada. La clase del objeto retornado
será un "htmlwidget" y dependiendo de la librería usada pertenecerá adicionalmente
a la clase "highchart", "plotly" o "dygraphs".
}
\description{
Esta función proporciona excelentes herramientas y opciones para la visualización
de series de tiempo dinámicas con el objetivo de estudiar la evolución de una o
varias variables a lo largo del tiempo. Dicha serie interactiva se puede representar
usando tres diferentes librerías que son \code{Highcharter}, \code{Plotly} y \code{Dygraph}, las
cuales usan internamente \code{JavaScript}.
}
\details{
Al usar el paquete \code{Highcharter} y usar las opciones de descarga, el nombre del
archivo descargado será la concatenación del plot graficado y la categoría usada,
así, por ejemplo, si se graficó la serie de tiempo para la categoría "Sede" el
nombre será \code{PlotSeries_Sede.png}.
}
\examples{
col <- c("#29ABE2", # AZUL CLARO  | Amazonia
         "#8CC63F", # VERDE       | Bogota
         "#CC241D", # ROJO        | Caribe
         "#FF3673", # MORADO CLARO| De la paz
         "#0071BC", # AZUL VIVO   | Manizales
         "#F15A24", # NARANJA     | Medellin
         "#FBB03B", # AMARILLO    | Orinoquia
         "#93278F", # MORADO      | Palmira
         "#8A381A") # GRIS        | Tumaco
Msj <- "Distribuci\u00f3n de estudiantes admitidos (desde el 2009-I al 2020-II) por sede."
Plot.Series(datos     = Consolidado,
            categoria = "SEDE_NOMBRE_ADM",
            colores   = col,
            titulo    = "EVOLUCI\u00d3N DEL N\u00daMERO DE GRADUADOS POR SEDE",
            labelY    = "N\u00famero de graduados (k: miles)",
            libreria  = "highcharter",
            estilo    = list(LegendTitle = "SEDE:", hc.Tema = 4, hc.Slider = TRUE, hc.Credits = Msj)
            )
Plot.Series(datos     = Consolidado,
            categoria = "SEDE_NOMBRE_ADM",
            colores   = col,
            titulo    = "EVOLUCI\u00d3N DEL N\u00daMERO DE GRADUADOS POR SEDE",
            labelY    = "N\u00famero de graduados (k: miles)",
            libreria  = "plotly",
            estilo    = list(LegendTitle = "SEDE:", ply.Interaction = "closest",
                             ply.LegendPosition = list(x = 0.16, y = -0.25, orientation = "h"),
                             ply.Credits = list(x = 0.5, y = 0.95, text = Msj))
            )
Plot.Series(datos     = Consolidado,
            categoria = "SEDE_NOMBRE_ADM",
            colores   = col,
            titulo    = "EVOLUCI\u00d3N DEL N\u00daMERO DE GRADUADOS POR SEDE",
            labelY    = "N\u00famero de graduados (k: miles)",
            libreria  = "dygraphs",
            estilo    = list(dyg.LegendWidth = 810, dyg.Resaltar = TRUE)
            )

}
